+++ CHANGES.md
@@ -1,4 +1,13 @@
 
+
+## v0.1.2 (2024-02-05)
+
+#### aslp_client_server_ocaml:
+
+- add offline lifter support
+- use int32 opcodes in server interface
+
+
 ## v0.1.1 (2025-02-04) 
 
 aslp_client_server_ocaml: Add in-memory cache on server side.